Revert "Fix 01-driver.t Darwin mocks for 691d510"
This reverts commit 040cd1b8cd5159f28c033f770c6ffa8f4166c59c.

I'm not sure whether the mock recordings will be properly
maintained. Since we only need them to match when a new CPAN version
is released, I'll remove it from the PR and pay attention to the
master branch to see if/when a new recording is needed. It makes more
sense just to make one set of recordings right beofre release instead
of trying to update distinct recordings for each PR when they'd just
get overwritten in master anyway.