Promote firefox_profile to init hash argument
I struggled with this for a few days - should the initiation look like

    my $fp_driver = Selenium::Remote::Driver->new(firefox_profile => $fp);

or

    my $fp_driver = Selenium::Remote::Driver->new(extra_capabilities => {
        firefox_profile => $fp
    });

My issue is I'm not sure what the 'extra' in 'extra_capabilities' really
refers to. This leads to, in practice, some confusion about what
arguments go where.

The disadvantage of having it as a primary arg is that it adds to the
confusion of what extra_capabilities is for in the first place. The
advantage is that it will fit into the future way that is mentioned in
issue #110 where everything in the init hash makes it into the
desiredCapabilities object in the eventual POST to /session.

It is probably internally consistent to have it in the
extra_capabilities hash considering the current implementation in 0.18,
but I was reluctant to do that because it wouldn't reflect the new API
that should supercede extra_capabilities.

I decided to promote it to a top level argument in light of the planned
changes from #110.