Add encoded profile to test fixtures
This should fix the failing tests.

When we encode_base64 the Firefox profile, it appears to generate a
slightly different string every single time. So, we're doing an ugly
thing where we encode it ourselves and sneak the pre-encoded profile
into desiredCapabilities by putting in extra_capabilities manually. It
means we're not testing the "transform from a ::Firefox::Profile to an
encoded thing" functionality in Driver.pm. But, we're still sure that
the encoded version works with the webdriver api when we record it, so
it has some use.

We don't need to re-encode a profile every time we run tests on
different platforms or anything like that - this one should be fine for
the forseeable future, so the toggle is just in Firefox-Profile.t and
not in an environment variable or anything.

[ci skip], since we haven't updated the linux mocks yet.