Move unregistered portion of driver.t testing to the bottom
The majority of the 01-driver.t test runs against a set of recorded HTTP
request/response pairs. To get a no server response message, we
unregister the PSGI takeover and get access to normal HTTP traffic. But,
we only want to unregister when we no longer need the recorded
pairs. The storage tests do need the recorded pairs, so that test should
go first; alternatively, this test should go last.