Ensure we skip_all when replaying with a missing mock_file
When refactoring the test harness setup into a separate lib, I made a
bug where OS types that we weren't prepared for would fail the tests
instead of skipping everything. CPAN testers found this by running a
test on cygwin windows.

cpan tester report on cygwin:
http://www.cpantesters.org/cpan/report/a87063d1-6e72-1014-bdc8-69e85cf4dae8

Because of the way the `t/lib/TestHarness.pm` was set up, we were never
getting to the `-e $mock_file` check that would use the `plan skip_all
=> ...` to allow the tests to pass when we don't have a mock
prepared. Instead, we were going straight into Mock::RemoteConnection
where it would rightly croak due to having non-existent file.

On the upswing, I found that we _can_ indeed do `plan skip_all => ...`
from TestHarness.pm instead of the `.t` file itself, so that's
nice. We're able to clean up the set up for each test down to one
command:

```perl
my %selenium_args = %{ TestHarness->new(
    this_file => $FindBin::Script
)->base_caps };
```

While it is pleasingly DRY, unfortunately, it's not very obvious what's
going on in there. That's not good. It's much more dry than the old
method, but none of that code indicates that we're deciding whether or
not to record, finding the mock files, etc, etc.