Fix TestHarness instantiation for DEMOLISH to fire properly
This was broken by f5b665c in #171.

Instantiating TestHarness, immediately calling one of its subroutines,
and de-referencing the returned hashref apparently also leads
TestHarness to immediately invoke DEMOLISH upon itself. Since we're
depending on the timing of DEMOLISH to be at the end of the .t file,
it was leading to empty mock recordings.

Separating the instantiation from the subroutine call and hash
de-referencing fixes the issue.