Add geolocation endpoints, although they're not widely implemented
I had trouble getting `set_geolocation` and `get_geolocation` to work
against any of the desktop browsers - I tried in Firefox, Chrome, and
Firefox with a few geolocation permissions tweaked in the
preferences. That's not too surprising, as they're really more useful on
mobile devices. Nevertheless, they're in now for use in Appium if
needed, or to be ready whenever the drivers end up implementing them.

For #113.