Fix whitespace in T:S:R:WebElement
It keeps showing up in the 'Provides' section on metacpan, instead of in
the Modules section. That gives this module unnecessary emphasis and may
be confusing if you don't know what you're looking for. Additionally, it
doesn't render the pod when a file is in the 'Provides' section, but
instead just shows the source and actually hides the pod. I'm not sure
if this will work, though...