Use Time::Mock to speed up waiter test execution
It's a bit unstable, so we need to watch the trial test reports to see
whether this is a reliable method. In order to even use it in the first
place, we opened up the intervals and expectations on the existing tests
significantly. This commit can be easily reverted; all it means is that
tests will take longer to run.