Fix & simplify Firefox extension disk layout
This was nice! We get to replace a non-core module (Archive::Extract)
with a core module (IO::Uncompress::Unzip) and instead of parsing XML
with regex, we're trying out XML::Simple. It's unfortunately not core,
but it doesn't have too many dependencies and hopefully our users
already have either or its deps already installed!

We weren't able to find any core XML modules :(