Refactor parameter resolution for upload file
This fixes a discrepancy between operating systems - on Windows,
Cwd::abs_path apparently dies if the file doesn't exist, whereas on OS X
it does not die. As a result, the tests were failing due to this
death.

Now, we only abs_path if the raw data isn't passed in.