Coerce timeouts in various set*timeout subroutines
The tests for these aren't perfect - that is, we only test the
successful coercion for `set_implicit_wait_timeout`, but not for all of
them. We _do_ test the failed coercion for all of the subroutines.

The reason is because I'm too lazy to re-record the recordings, and
refraining from adding additional successful set*timeout calls means we
don't make any additional requests to the server, so we don't need to
change the recordings.

In theory, this shouldn't cause any issues...