Give up if we're in gd in CanStartBinary
There's no guarantee that we can do anything if we're global
destruction. The only reasonable thing to do is to bail out so as not to
throw any warnings or errors; to guarantee a cleanup, the user _must_
make sure to dispose of their objects appropriately.

Specifically in regards to CanStartBinary, the `system` process that we
created earlier to run the binary does seem to kill itself before global
destruction, so that's nice.