Extract install.rdf from Firefox extensions manually
With approximately zero knowledge of how zip files work, our issue was
that the `install.rdf` that we needed from the XPI file was not listed
in the manifest that IO::Uncompress::Unzip::unzip was reading. This
prohibits us from using the lovely `Name => 'install.rdf'` shortcut that
was so handy. As a result, we must traverse the zip manually, but in the
end the results are the same: the `em:id` is extracted from the xpi.