Fix find_child_element() and find_child_elements() not working with default_finder
When not explicitly passing a locator scheme to find_child_element(s),
default_finder was erronously looked up in FINDERS again. This just
happens to work for finders like xpath that are called the same in either
case but breaks for e.g. css which maps to "css selector".
Fixed by using the same correct code as in find_element() and find_elements().