Find child element default finder fix (#270)
* Stop suggesting linux help when making recordings

The docker image in gempesaw/docker-selenium-remote-driver takes care of
everything for us.

* Tweak error handler test in 01-driver.t

It wasn't calling the error handler for a find_element, oddly. I can't
recall if this is expected behavior ?

* Fix find_child_element() and find_child_elements() not working with default_finder

When not explicitly passing a locator scheme to find_child_element(s),
default_finder was erronously looked up in FINDERS again. This just
happens to work for finders like xpath that are called the same in either
case but breaks for e.g. css which maps to "css selector".
Fixed by using the same correct code as in find_element() and find_elements().

* Add test to exercise default finder bug from #256

* Update os x bindings for #256

* Update linux bindings for #256

Fix #256