Fix browser initialization
Since Selenium IDE is a firefox plugin I think we can asume
there is a firefox installed on the machine this script was
generated. This is why I hardcoded firefox startup here.

Neithertheless we have to pick a browser here, since asking
the user which browser to use is currently not possible via the
plugin structure