Look for session id where geckodriver puts it
When creating a new session, the standalone server responds with

    { sessionId => ..., }

Looks like the geckodriver server's response is like

    { value => { sessionId => ... } }

So...it seems like we have to put the sessionId there manually in the
geckodriver case. Curiously, the w3c spec does indicate that sessionId
should be a top level property in the json response, but...whatever.