Fix #345: Implement WC3 webdriver standard compatibility
This adds a new module, Selenium::Remote::Spec that describes
the new specification, along with processing methods to simplify
shims needed in multiple methods.

Current Firefox, Chrome and Edge are verified as working, and
have a variety of tweaks to ensure as many methods as possible
work as well (or better) than they previously did.

More work will need to be done in the future to complete the
JSONWire protocol, so we can both fall back when legacy methods
are available and also be more useful to users in general.

An acceptance testing directory (at) has been added with tests
designed to be run against the various drivers directly by
module contributors and authors.  The idea here is that we can
keep an eye on what works and does not via usage of these tests.