Fix #386: Adjust SRD::WE::is_displayed for new "displayedness"
This is sort of the "why" behind the extreme rustling that happened
when an element bounding box check was added to firefox JSONWIRE
driver, but the is_displayed method was not updated.

It also didn't hurt that their BB calc was wrong when you used
negative margins, but that's neither here nor there.  WC3 & JS
polyfill calculations are correct now.