fix search for unix binary on MacOS
On MacOS, the `lsof -i :<port>` command seems to show multiple processes
have the port open, one of which is the current Perl program. Since this
one mostly comes first in the output, `shutdown_unix_binary` tries to
kill the `perl` process, not the browser process. Filtering out the
current `perl` process from the process we're trying to kill fixes this
issue.